Shift legacy NumPy distributions to NumPy module
